добавил компонент place-card и тест для него
